struct-storage works; tests separated, de-aliased.
